backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-aiplatform
pydantic
pydantic-settings
gunicorn
google-cloud-discoveryengine